+++ CHANGELOG.md
@@ -6,6 +6,7 @@ dd mm yyyy
 
 - **BC Break**: Updated [Neucore plugin](https://github.com/tkhamez/neucore-plugin) to 0.10.0.
 - Added search for service accounts to *Member Data -> Characters*.
+- Duplicate character search results have been removed.
 
 ## 1.41.0